client: Refactors models->serializers
Because:
- It was confusing have a folder named models but only have a serializer
  in it
