+++ webclient/src/hooks/useRecords.js
@@ -1,6 +1,5 @@
 import React, { useCallback, useState, useEffect } from "react";
 import { useRecordRepository } from "../hooks/useRepository";
-import { REC_HASH_RATE } from "../models/Fixture";
 
 export const useRecords = ({ labelId, state, tick }) => {
   const [records, setRecords] = useState(null);
