+++ webclient/src/repos/ERC721Repository.js
@@ -1,6 +1,6 @@
 import client from "./ApiClient";
 import Base from "./Base";
-import { ERC721sDeserializer } from "./models/ERC721.js";
+import { ERC721sDeserializer } from "./serializers/ERC721.js";
 
 // ERC721Repository is the data access interface  for labels
 class ERC721Repository extends Base {
