+++ webclient/src/repos/serializers/Member.js
@@ -2,7 +2,7 @@ export const MembersDeserializer = (data) => {
   return {
     id: data.id,
     name: data.name,
-    wallet_address: data.wallet_address,
+    walletAddress: data.wallet_address,
     labelId: data.recordlabel,
   };
 };
